Added benchmarks and tests
