+++ hw10_program_optimization/benchmark_results/benchmark_3.txt
@@ -1,0 +1,7 @@
+goos: linux
+goarch: amd64
+pkg: github.com/murashko91/otus-go/hw10_program_optimization
+cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
+BenchmarkGetDomainStat-4   	     123	   9763617 ns/op	   19664 B/op	      83 allocs/op
+PASS
+ok  	github.com/murashko91/otus-go/hw10_program_optimization	2.225s
