+++ hw10_program_optimization/benchmark_results/benchstat_1.txt
@@ -1,0 +1,21 @@
+goos: linux
+goarch: amd64
+pkg: github.com/murashko91/otus-go/hw10_program_optimization
+cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
+                │ benchmark_0.txt │         benchmark_1.txt         │
+                │     sec/op      │    sec/op     vs base           │
+GetDomainStat-4    173.989m ± ∞ ¹   7.989m ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
+
+                │ benchmark_0.txt  │         benchmark_1.txt          │
+                │       B/op       │     B/op       vs base           │
+GetDomainStat-4   132975.1Ki ± ∞ ¹   124.0Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
+
+                │ benchmark_0.txt │        benchmark_1.txt         │
+                │    allocs/op    │  allocs/op   vs base           │
+GetDomainStat-4   1700495.0 ± ∞ ¹   411.0 ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
