+++ hw10_program_optimization/benchmark_results/benchstat_3.txt
@@ -1,0 +1,21 @@
+goos: linux
+goarch: amd64
+pkg: github.com/murashko91/otus-go/hw10_program_optimization
+cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
+                │ benchmark_2.txt │         benchmark_3.txt         │
+                │     sec/op      │    sec/op     vs base           │
+GetDomainStat-4     11.765m ± ∞ ¹   9.764m ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
+
+                │ benchmark_2.txt │         benchmark_3.txt          │
+                │      B/op       │     B/op       vs base           │
+GetDomainStat-4     39.57Ki ± ∞ ¹   19.20Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
+
+                │ benchmark_2.txt │        benchmark_3.txt         │
+                │    allocs/op    │  allocs/op   vs base           │
+GetDomainStat-4      320.00 ± ∞ ¹   83.00 ± ∞ ¹  ~ (p=1.000 n=1) ²
+¹ need >= 6 samples for confidence interval at level 0.95
+² need >= 4 samples to detect a difference at alpha level 0.05
