+++ hw10_program_optimization/stats_test.go
@@ -1,3 +1,4 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
@@ -35,5 +36,12 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+
+	t.Run("Unexpected JSON error", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString("Unexpected JSON"), "unknown")
+		var expResult DomainStat
+		require.Error(t, err)
+		require.Equal(t, expResult, result)
 	})
 }
